backend/requirements.txt
fastapi==0.116.1
openpyxl==3.1.5
pyproj==3.7.1
uvicorn[standard]==0.40.0
python-multipart==0.0.20
numpy==2.2.4
pandas==2.2.3
scipy==1.15.2
scikit-learn==1.6.1
xgboost==3.0.0
PyKrige==1.7.2
matplotlib==3.10.1
reportlab==4.3.1
python-docx==1.1.2
python-pptx==1.0.2
Pillow==11.1.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-cloud-pubsub==2.29.0
google-cloud-logging==3.12.1
google-cloud-run==0.15.0
google-cloud-secret-manager==2.23.2
anthropic[vertex]==0.61.0
